gcs: Retry API calls.
In our experience, GCS API calls fail quite regularily, e.g. once every
few thousand operations. The python API client that we are using
already supports retries with exponential backoffs, so that is
what we are implementing here. The API is a bit clunky in that
we can't set it globally for all requests but have to do it at
every call site though.